Refactor in whole
