+++ src/models/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './Webhook';
+export * from './User';
