+++ src/routers/index.ts
@@ -1,0 +1,5 @@
+import control from './control';
+import directapi from './direct-api';
+import webhooks from './webhooks';
+
+export { control, directapi, webhooks };
